Create application integrations.
The first commit brought this project most of its actual tool framework,
but this commit integrates the bulk of it.

* Router wraps React Router Dom (with a custom Browser History)
* Provider wraps React Redux (with Redux Sagas)
* Shell provides both of those things to whatever you want inside.
* App then provides the gateway to all of that.

So here's the thing.  Originally this whole situation was huge.  I had
components, elements (I was calling styled components elements), and
screens.  I animated the screens.  It was great.

But, I had no idea what I was doing, so I eradicated all of that just a
bit ago and I'm only including the barest amount right now.  Here's what
I'd like to do instead:

1. I'm probably going to just create a nice layout.
2. I will install chroma js, but not use it much.
3. I will start doing things with my UI reducers.
4. I will make a Drawer, a Menu, and a Modal system.

That will be the meat of my application shell, in the end.  Other
things, they will use those components (or necessitate a new component,
maybe).  That's what I'll call my first milestone for this.

The next will be actual posting, then editing, then we'll go from there.