Add UI behavior and components.
This commit introduces some essential chrome to the App, and is probably
the last pre-logic setup commit.

The three new components are:

* `<Modal />`
* `<Drawer />`
* `<Menu />`

And they serve more or less as visually significant ways to offer
different types of interactions to a user.  They each accept children
and a visibility toggle, and they're each wired into the `ui` slice of
the redux state tree.

Two questions stick out in my head:

* There's a lot of repetition in the styles.  How can I reduce this in a
  way which is still pleasant to maintain going forward?
* Should the Modal / Drawer / Menu elements be put into the App purview?
  Is App really a marriage of Bootstrapping + Chrome?