Create Note core domain.
This commit introduces the note core domain and prepares to use it in
tandem with Draft.js (through the Editor component).

To explain myself, the Editor component came first and it was pretty
nice (to a certain degree). Then I'd realized I was spiking things out
while reading the Draft js docs.  Not good.

So, in comes the Core domain.  This domain is ported from a previous
work (see esmevane/journaling for more) where I was more back-end
focused.  But, that back end architecture is important here, because in
the land of React / Redux, you are constantly coerced into a
component-centric design.

Here's the point:

From here on out, most features will be implemented through this core
domain, and exposed through some sort of boundary built deliberately on
top of that.  Then Redux will provide that behavior to the component
layer.

For example, persistence (thinking of GunDB here) will be a Storage
strategy implemented through Core Storage.

As another example, Draft JS will be integrated first with Core
(possibly through a new bounded context) and then after that's done,
that context (not Draft JS itself) will be integrated with the Editor.