Add Greeting component and integration tests.
I may move them later, but for now I've walked a lot of the work I'd
done so far back and instead started to implement pieces of it story by
story.

This commit removes the Editor (which I'd put there with no tests or
stories whatsoever) and replaces it with a prompt / greeting component
which I might be using to invite folks to edit pages.

I've also put my Redux and Router integrations under a small integration
test, but this precipitates some integration tests I anticipate needing
next:

1. I will need to access the store directly in order to observe / test
   Redux and Redux Saga integrations.

2. I will need to access the history directly in order to test
   route-based screens and features.