Add an animated editable page switcher.
This commit took me a while, but it was worth it.  Animation is a
feature, I don't care who promises you it's doable quick and easily, it
generally only fits that description in scenarios where people don't
want to test or structure it correctly.

Harumph.

Anyway, what we have here is a Switch component inheritor which provides
its utility alongside an animated transition effect.  So anywhere you
want to have mutually exclusive components appear and vanish in a nice
upward slide, you can wrap them in this switcher.

One issue with this is that it relies on absolute positioning to avoid a
visual glitch as the divs appear / leave the DOM, but it doesn't control
this styles itself.  That needs to be fixed, but I'm shipping this as I
didn't want to rabbit hole on that issue just yet.