Fall cleaning.
This commit removes a lot.  The summary is this:

* Slide Switch is gone.
* Markdown dreams, dead.
* Core notes are gone.
* Integration testing, gone.
* Animations in general, gone.

Not using Overpass Mono anymore.  But I'm keeping it.

How can I explain the thinking behind this commit?  Basically, I had
ambitions, and I wrote them into the software, but the software didn't
do anything, really.  It just described some ambitions.  And then I used
those ambitions as an excuse to get sidetracked.

No more of that.

We've also got a substantial clean-up of a good amount of the remaining
code.  It'll be in less complex organizations and hopefully in a few
more minutes I'll have my redux state organized, as well.