Allow editing of the index page.
This is probably a pretty weak update, but it establishes a few concepts
I've been kicking around in my head.

* State is now in the src/state folder.
* The editor is now persisting things via redux.

Honestly the biggest thing I've proven here is that stripping a lot of
custom things away is a good thing.  So, in an upcoming commit, I'll
probably be stripping more out and making it even simpler.

Another thing I'll be doing is structuring sagas and services first,
then wiring those up to the component layer, which will just be a
declarative portrayal of the service strata.