Integrate repos with redux saga.
Okay, just ONE repo is integrated with redux saga.  But it's a pretty
passthrough style integration, one which ought to allow me to plug in
more entities as time goes on.

Concretely, this basically means that we've got the async Note
repository, and redux saga now observes an instance of that and passes
its emitted events through to redux in general.