Add PouchDB support.
GunDB has to go, let's start with that.  It sounds pretty great in
abstract (GunDB, I mean) - but when you zero in on what's really going
on in the life of using it, things aren't as great.

Let's start:  It's clearly a commercial product that has open source
offering as a "try it before you buy" option.  It's pesterware.  And
finally, it's got some not-so-useful errors.  So, this commit introduces
PouchDB instead.

I have more gripes, but let's move on.

PouchDB is a step down from GunDB in terms of project size, but it seems
to have a sincere open source presence, and gives most of the same
features, sometimes with a little bit of a verbose API.  There's also a
straightforward replication and distribution strategy for PouchDB, which
is nice.

Look forward to a GunDB removal commit, coming to a repo near you.