Reorganize things.
It never sat well with me to put everything in a single folder.  This is
an insubstantial change except that it refactors a lot of our ... stuff
to be more discoverable.

* Is it a visual thing?  It should go in Elements.
* Is it a layout thing?  It should go in Layout.
* Is it a logical thing?  It should go in Components.
* Is an entire page of things?  It should go in Screens.

So there are some rules for each, but in a strict sense, elements and
layout are kind of loosey-goosey right now.

I'd like to change a few things to wrap them with `createX` function
calls.  It's cheap right now to implement this change and it lets me do
things like create measurement systems or user preference systems and
pass those in a "Config" object through these functions.  Then, the
functions can return different things depending on those configurations.